fixing schema, domain and mapping

+++ WeCare.Domain/Models/ParticipationCertificate.cs
@@ -2,12 +2,12 @@ namespace WeCare.Domain;
 
 public class ParticipationCertificate
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public DateOnly CreationDate { get; set; }
 
     public int CandidateId { get; set; }
     public Candidate Candidate { get; set; }
 
-    public int VolunteerOpportunityId { get; set; }
+    public long VolunteerOpportunityId { get; set; }
     public VolunteerOpportunity VolunteerOpportunity { get; set; }
 }
